pyproject.toml
[tool.poetry]
name = "maya-sawa-v2"
version = "0.1.0"
description = "Behold My Awesome Project!"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "maya_sawa_v2"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1.11"
django-environ = "0.12.0"
django-model-utils = "5.0.0"
django-allauth = {extras = ["mfa"], version = "65.10.0"}
fido2 = "1.1.2"
django-crispy-forms = "2.4"
crispy-bootstrap5 = "2025.6"
django-redis = "6.0.0"
whitenoise = "6.9.0"
redis = "6.2.0"
hiredis = "3.2.1"
uvicorn = {extras = ["standard"], version = "0.35.0"}
uvicorn-worker = "0.3.0"
python-slugify = "8.0.4"
Pillow = "11.3.0"
argon2-cffi = "25.1.0"
djangorestframework = "3.15.2"
celery = "5.4.0"
openai = "1.12.0"
google-generativeai = "0.8.3"
dashscope = "1.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.4.1"
pytest-django = "4.11.1"
pytest-sugar = "1.0.0"
mypy = "1.17.1"
django-stubs = {extras = ["compatible-mypy"], version = "5.2.2"}
ruff = "0.12.7"
coverage = "7.10.1"
djlint = "1.36.4"
pre-commit = "4.2.0"
factory-boy = "3.3.2"
django-debug-toolbar = "6.0.0"
django-extensions = "4.1"
django-coverage-plugin = "3.1.1"
Werkzeug = {extras = ["watchdog"], version = "3.1.3"}
ipdb = "0.13.13"
watchfiles = "1.1.0"
sphinx = "8.2.3"
sphinx-autobuild = "2024.10.3"
drf-spectacular = "0.28.0"

[tool.poetry.group.prod.dependencies]
gunicorn = "23.0.0"
psycopg = "3.2.9"
django-anymail = {extras = ["mailgun"], version = "13.0.1"}

[tool.poetry.scripts]
start = "manage:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["maya_sawa_v2/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TC",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
  "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true
